admin routes

+++ controllers/workerControllers.js
@@ -6,7 +6,7 @@ const workerControllers = {
         try {
             const workerExist = await Worker.findOne({ email })
             if (workerExist) {
-                res.json({ success: false, error: "The email user is already in use", response: null })
+                res.json({ success: false, error: "The user email is already in use", response: null })
 
             } else {
                 const newWorker = new Worker({
@@ -19,8 +19,26 @@ const workerControllers = {
                     schedule,
                     reviews
                 })
-                await newWorker.save()
-                res.json({ success: true, response: { newWorker }, error: null })
+                if (req.user.admin) {
+                    await newWorker.save()
+                    res.json({ success: true, response: { newWorker }, error: null })
+                } else {
+                    res.json({ success: false, error: 'Unauthorized User, you must be an admin' })
+                }
+            }
+        } catch (error) {
+            res.json({ success: false, response: null, error: error })
+        }
+    },
+    getWorkers: async (req, res) => {
+
+        try {
+            if (req.user.admin) {
+
+                const users = await Worker.find()
+                res.json({ success: true, users })
+            } else {
+                res.json({ success: false, error: 'Unauthorized User, you must be admin' })
             }
         } catch (error) {
             res.json({ success: false, response: null, error: error })
@@ -28,24 +46,35 @@ const workerControllers = {
     },
     deleteWorker: async (req, res) => {
         const id = req.params.id
-        let worker
         try {
-            await Worker.findOneAndDelete({ _id: id })
+            if (req.user.admin) {
+                await Worker.findOneAndDelete({ _id: id })
+                res.json({ success: true, msg: 'Worker deleted successfully' })
+
+            } else {
+                res.json({ success: false, error: 'Unauthorized User, you must be an admin' })
+            }
         } catch (error) {
             console.log(error)
         }
-        res.json({ respuesta: worker })
     },
     modifyAWorker: async (req, res) => {
         let id = req.params.id
         let worker = req.body
         let update
         try {
-            update = await Worker.findOneAndUpdate({ _id: id }, worker, { new: true })
+            if (req.user.admin) {
+
+                update = await Worker.findOneAndUpdate({ _id: id }, worker, { new: true })
+                res.json({ success: true, update })
+            } else {
+                res.json({ success: false, error: 'Unauthorized User, you must be an admin' })
+            }
         } catch (error) {
             console.log(error)
+            res.json({ success: false, error })
         }
-        res.json({ success: update ? true : false })
     },
+
 };
 module.exports = workerControllers;
